Updated for Scala 3
