+++ .scalafmt.conf
@@ -1,10 +1,9 @@
-version = 2.7.5
-
-style = default
-
+version = 3.5.9
 maxColumn = 100
 
 align.preset = none
+
+runner.dialect = scala3
 
 rewrite.rules = [
   AvoidInfix
