+++ build.sbt
@@ -1,17 +1,19 @@
+import sbt.librarymanagement.CrossVersion
+
+lazy val scala2 = "2.13.14"
+lazy val scala3 = "3.5.1"
+lazy val supportedScalaVersions = List(scala2, scala3)
+
+
 ThisBuild / organization := "simex"
 
-ThisBuild / version := "0.8.0" // Keep the version in sync with simex-messaging
+ThisBuild / version := "0.9.0"
 
 lazy val commonSettings = Seq(
-  scalaVersion := "2.13.10",
+  scalaVersion := scala3,
   libraryDependencies ++= Dependencies.all,
-  resolvers += Resolver.githubPackages("TheDiscProg"),
   githubOwner := "TheDiscProg",
-  githubRepository := "simex-rabbitmq",
-  addCompilerPlugin(
-    ("org.typelevel" %% "kind-projector" % "0.13.2").cross(CrossVersion.full)
-  ),
-  addCompilerPlugin("com.olegpy" %% "better-monadic-for" % "0.3.1")
+  githubRepository := "simex-rabbitmq"
 )
 
 lazy val root = project.in(file("."))
@@ -21,7 +23,17 @@ lazy val root = project.in(file("."))
   .settings(
     commonSettings,
     name := "simex-rabbitmq",
-    scalacOptions ++= Scalac.options
+    scalacOptions ++=Scalac.options,
+    scalacOptions ++= {
+      CrossVersion.partialVersion(scalaVersion.value) match {
+        case Some((2,13)) => Seq("-Ytasty-reader")
+        case _ => Seq("-Yretain-trees")
+      }
+    },
+    crossScalaVersions := supportedScalaVersions,
+    libraryDependencies ++= Seq(
+      ("com.github.pureconfig" %% "pureconfig" % "0.17.7").cross(CrossVersion.for3Use2_13)
+    )
   )
 
 lazy val integrationTest = (project in file("it"))
@@ -31,6 +43,18 @@ lazy val integrationTest = (project in file("it"))
     name := "simex-rabbitmq-integration-test",
     publish / skip := true,
     libraryDependencies ++= Dependencies.it,
+    libraryDependencies ++= Seq(
+      ("com.dimafeng" %% "testcontainers-scala-scalatest" % "0.41.4").cross(CrossVersion.for3Use2_13)
+    ),
+    libraryDependencies ++= {
+      if (scalaVersion.value.startsWith("2"))
+        Seq(
+          compilerPlugin(("org.typelevel" %% "kind-projector" % "0.13.3").cross(CrossVersion.full)),
+          compilerPlugin(("com.olegpy" %% "better-monadic-for" % "0.3.1")),
+        )
+      else
+        Seq()
+    },
     parallelExecution := false,
     coverageFailOnMinimum := true,
     coverageMinimumStmtTotal := 85,
