+++ src/main/scala/io/github/thediscprog/simexrabbitmq/consumer/SimexMQConsumerAlgebra.scala
@@ -1,0 +1,8 @@
+package io.github.thediscprog.simexrabbitmq.consumer
+
+import fs2.Stream
+
+trait SimexMQConsumerAlgebra[F[_]] {
+
+  def consumeRMQSimexMessage(handlers: List[SimexMessageHandler[F]]): Stream[F, Unit]
+}
