+++ src/main/scala/io/github/thediscprog/simexrabbitmq/consumer/SimexMessageHandler.scala
@@ -1,0 +1,9 @@
+package io.github.thediscprog.simexrabbitmq.consumer
+
+import io.github.thediscprog.simexmessaging.messaging.Simex
+import io.github.thediscprog.simexrabbitmq.RabbitQueue
+
+case class SimexMessageHandler[F[_]](
+    queue: RabbitQueue,
+    f: Simex => F[Unit]
+)
