--- src/main/scala/simex/rabbitmq/RabbitQueue.scala
@@ -1,13 +1,0 @@
-package simex.rabbitmq
-
-import dev.profunktor.fs2rabbit.model.{ExchangeName, ExchangeType, QueueName, RoutingKey}
-
-trait RabbitQueue {
-  val name: QueueName
-  val routingKey: RoutingKey
-  val exchange: ExchangeName
-  val dlx: Option[String]
-  val exchangeType: ExchangeType
-  val messageTTL: Option[Long]
-  val consumers: Boolean
-}
