--- src/main/scala/simex/rabbitmq/consumer/SimexMQConsumerAlgebra.scala
@@ -1,8 +1,0 @@
-package simex.rabbitmq.consumer
-
-import fs2.Stream
-
-trait SimexMQConsumerAlgebra[F[_]] {
-
-  def consumeRMQSimexMessage(handlers: List[SimexMessageHandler[F]]): Stream[F, Unit]
-}
