--- src/main/scala/simex/rabbitmq/consumer/SimexMessageHandler.scala
@@ -1,9 +1,0 @@
-package simex.rabbitmq.consumer
-
-import simex.messaging.Simex
-import simex.rabbitmq.RabbitQueue
-
-case class SimexMessageHandler[F[_]](
-    queue: RabbitQueue,
-    f: Simex => F[Unit]
-)
