--- src/main/scala/simex/rabbitmq/entities/ServiceError.scala
@@ -1,14 +1,0 @@
-package simex.rabbitmq.entities
-
-import enumeratum.{Enum, EnumEntry}
-
-sealed trait ServiceError extends EnumEntry {
-  val message: String
-}
-
-case object ServiceError extends Enum[ServiceError] {
-
-  case class InvalidMessageFormat(message: String) extends ServiceError
-
-  override def values: IndexedSeq[ServiceError] = findValues
-}
