--- src/main/scala/simex/rabbitmq/publisher/SimexMQPublisherAlgebra.scala
@@ -1,9 +1,0 @@
-package simex.rabbitmq.publisher
-
-import simex.messaging.Simex
-import simex.rabbitmq.RabbitQueue
-
-trait SimexMQPublisherAlgebra[F[_]] {
-
-  def publishMessageToQueue(message: Simex, queue: RabbitQueue): F[Unit]
-}
